fix task dependencies
The source code for `DAGTask` says the following (abbreviated):

```python
    def __rshift__(self, other):
        """Implement task1 >> task2."""
        # ...
        tasks = _convert_func_to_task(other)
        self.add_successors(tasks)
        return tasks
```

So the `>>` operator seems to work in the opposite way to what the script says. `a >> b` means that `b` will run after `a`, not before.

This is also consistent with the [Airflow API](https://airflow.apache.org/docs/apache-airflow/stable/core-concepts/dags.html#task-dependencies).